- New constructor that throws exceptions is added - Extract method refactoring applied to reduce duplicate code
Calm down. The interface is not changed beyond the addition of
the new constructor and its signature is different enough so it
will not be mistaken. Also the new refactored init method is 
in use _only_ in the newly created constructor so no previous
code has undergone surgery. Obviously. The other constructors 
would benefit of using it though, if only for the sake of
maintenance.




git-svn-id: http://multitude.googlecode.com/svn/trunk@398 d2709d7f-9d48-0410-97af-374aaeb070bb